Separate libs file: '.h' to '.c' and '.h'
